chore: rename `keyring2` to `keyring`
