+++ packages/snap/snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/MetaMask/template-snap-monorepo.git"
   },
   "source": {
-    "shasum": "0HnSdy81znkETMJMXhZF4w5HawtGsGPbqZi8bgdmaJI=",
+    "shasum": "ffLL4S8PhSjED4b9HYngreUaE7yoF3rSPteI7I/V48o=",
     "location": {
       "npm": {
         "filePath": "dist/bundle.js",
