+++ packages/snap/src/util.ts
@@ -1,7 +1,7 @@
 import { JsonTx } from '@ethereumjs/tx';
 import type { Json } from '@metamask/utils';
 
-import { Wallet } from './keyring2';
+import { Wallet } from './keyring';
 import {
   InternalMethod,
   RequestMethods,
